+++ src/main/java/mdteam/ait/tardis/Tardis.java
@@ -318,15 +318,14 @@ public class Tardis {
      */
     public void tick(MinecraftClient client) { // fixme should likely be in ClientTardis instead, same with  other server-only things should be in ServerTardis
         // referencing client stuff where it COULD be server causes problems
-        if(client.player != null &&
-                ClientTardisUtil.isPlayerInATardis() && ClientTardisUtil.getCurrentTardis().equals(this) &&
-                this.getTravel() != null && this.getTravel().getState() != TardisTravel.State.LANDED) {
-            /*if (ClientShakeUtil.shouldShake(this)) */
+        if(ClientShakeUtil.shouldShake(this)) {
             ClientShakeUtil.shakeFromConsole();
         }
 
-        ClientTardisUtil.tickPowerDelta();
-        ClientTardisUtil.tickAlarmDelta();
+        if (this.equals(ClientTardisUtil.getCurrentTardis())) {
+            ClientTardisUtil.tickPowerDelta();
+            ClientTardisUtil.tickAlarmDelta();
+        }
     }
 
     public boolean isDirty() {
@@ -351,7 +350,13 @@ public class Tardis {
         }
     }
 
+    // todo move these up
+
     public boolean isInDanger() {
         return this.getHandlers().getHADS().isInDanger();
+    }
+
+    public boolean inFlight() {
+        return this.getTravel().getState() != TardisTravel.State.LANDED;
     }
 }