+++ src/main/java/mdteam/ait/tardis/exterior/variant/box/PoliceBoxSoulVariant.java
@@ -1,4 +1,4 @@
-package mdteam.ait.tardis.exterior.variant;
+package mdteam.ait.tardis.exterior.variant.box;
 
 public class PoliceBoxSoulVariant extends PoliceBoxVariant {
     public PoliceBoxSoulVariant() {